Cleaning up (#1213)
