+++ test/python/circuit/test_circuit_operations.py
@@ -19,84 +19,11 @@ from qiskit import QuantumRegister, ClassicalRegister, QuantumCircuit
 from qiskit import execute
 from qiskit import QISKitError
 from qiskit.quantum_info.states import state_fidelity
-from .common import QiskitTestCase
+from ..common import QiskitTestCase
 
 
-class TestCircuit(QiskitTestCase):
-    """QuantumCircuit basic tests."""
-
-    def test_qregs(self):
-        """Test getting quantum registers from circuit.
-        """
-        qr1 = QuantumRegister(1)
-        qr2 = QuantumRegister(2)
-        qc = QuantumCircuit(qr1, qr2)
-        q_regs = qc.qregs
-        self.assertEqual(len(q_regs), 2)
-        self.assertEqual(q_regs[qr1.name], qr1)
-        self.assertEqual(q_regs[qr2.name], qr2)
-
-    def test_cregs(self):
-        """Test getting classical registers from circuit.
-        """
-        cr1 = ClassicalRegister(1)
-        cr2 = ClassicalRegister(2)
-        cr3 = ClassicalRegister(3)
-        qc = QuantumCircuit(cr1, cr2, cr3)
-        c_regs = qc.cregs
-        self.assertEqual(len(c_regs), 3)
-        self.assertEqual(c_regs[cr1.name], cr1)
-        self.assertEqual(c_regs[cr2.name], cr2)
-
-    def test_circuit_qasm(self):
-        """Test circuit qasm() method.
-        """
-        qr1 = QuantumRegister(1, 'qr1')
-        qr2 = QuantumRegister(2, 'qr2')
-        cr = ClassicalRegister(3, 'cr')
-        qc = QuantumCircuit(qr1, qr2, cr)
-        qc.u1(0.3, qr1[0])
-        qc.u2(0.2, 0.1, qr2[0])
-        qc.u3(0.3, 0.2, 0.1, qr2[1])
-        qc.s(qr2[1])
-        qc.s(qr2[1]).inverse()
-        qc.cx(qr1[0], qr2[1])
-        qc.barrier(qr2)
-        qc.cx(qr2[1], qr1[0])
-        qc.h(qr2[1])
-        qc.x(qr2[1]).c_if(cr, 0)
-        qc.y(qr1[0]).c_if(cr, 1)
-        qc.z(qr1[0]).c_if(cr, 2)
-        qc.barrier(qr1, qr2)
-        qc.measure(qr1[0], cr[0])
-        qc.measure(qr2[0], cr[1])
-        qc.measure(qr2[1], cr[2])
-        expected_qasm = """OPENQASM 2.0;
-include "qelib1.inc";
-qreg qr1[1];
-qreg qr2[2];
-creg cr[3];
-u1(0.300000000000000) qr1[0];
-u2(0.200000000000000,0.100000000000000) qr2[0];
-u3(0.300000000000000,0.200000000000000,0.100000000000000) qr2[1];
-s qr2[1];
-sdg qr2[1];
-cx qr1[0],qr2[1];
-barrier qr2[0],qr2[1];
-cx qr2[1],qr1[0];
-h qr2[1];
-if(cr==0) x qr2[1];
-if(cr==1) y qr1[0];
-if(cr==2) z qr1[0];
-barrier qr1[0],qr2[0],qr2[1];
-measure qr1[0] -> cr[0];
-measure qr2[0] -> cr[1];
-measure qr2[1] -> cr[2];\n"""
-        self.assertEqual(qc.qasm(), expected_qasm)
-
-
-class TestCircuitCombineExtend(QiskitTestCase):
-    """Test combining and extending of QuantumCircuits."""
+class TestCircuitOperations(QiskitTestCase):
+    """QuantumCircuit Operations tests."""
 
     def test_combine_circuit_common(self):
         """Test combining two circuits with same registers.
