+++ test/python/circuit/test_circuit_qasm.py
@@ -1,0 +1,72 @@
+# -*- coding: utf-8 -*-
+
+# Copyright 2018, IBM.
+#
+# This source code is licensed under the Apache License, Version 2.0 found in
+# the LICENSE.txt file in the root directory of this source tree.
+
+# pylint: disable=unused-import
+
+"""Test Qiskit's QuantumCircuit class."""
+
+import os
+import tempfile
+import unittest
+
+import qiskit.extensions.simulator
+from qiskit import Aer
+from qiskit import QuantumRegister, ClassicalRegister, QuantumCircuit
+from qiskit import execute
+from qiskit import QISKitError
+from qiskit.quantum_info.states import state_fidelity
+from ..common import QiskitTestCase
+
+
+class TestCircuitQasm(QiskitTestCase):
+    """QuantumCircuit Qasm tests."""
+
+    def test_circuit_qasm(self):
+        """Test circuit qasm() method.
+        """
+        qr1 = QuantumRegister(1, 'qr1')
+        qr2 = QuantumRegister(2, 'qr2')
+        cr = ClassicalRegister(3, 'cr')
+        qc = QuantumCircuit(qr1, qr2, cr)
+        qc.u1(0.3, qr1[0])
+        qc.u2(0.2, 0.1, qr2[0])
+        qc.u3(0.3, 0.2, 0.1, qr2[1])
+        qc.s(qr2[1])
+        qc.s(qr2[1]).inverse()
+        qc.cx(qr1[0], qr2[1])
+        qc.barrier(qr2)
+        qc.cx(qr2[1], qr1[0])
+        qc.h(qr2[1])
+        qc.x(qr2[1]).c_if(cr, 0)
+        qc.y(qr1[0]).c_if(cr, 1)
+        qc.z(qr1[0]).c_if(cr, 2)
+        qc.barrier(qr1, qr2)
+        qc.measure(qr1[0], cr[0])
+        qc.measure(qr2[0], cr[1])
+        qc.measure(qr2[1], cr[2])
+        expected_qasm = """OPENQASM 2.0;
+include "qelib1.inc";
+qreg qr1[1];
+qreg qr2[2];
+creg cr[3];
+u1(0.300000000000000) qr1[0];
+u2(0.200000000000000,0.100000000000000) qr2[0];
+u3(0.300000000000000,0.200000000000000,0.100000000000000) qr2[1];
+s qr2[1];
+sdg qr2[1];
+cx qr1[0],qr2[1];
+barrier qr2[0],qr2[1];
+cx qr2[1],qr1[0];
+h qr2[1];
+if(cr==0) x qr2[1];
+if(cr==1) y qr1[0];
+if(cr==2) z qr1[0];
+barrier qr1[0],qr2[0],qr2[1];
+measure qr1[0] -> cr[0];
+measure qr2[0] -> cr[1];
+measure qr2[1] -> cr[2];\n"""
+        self.assertEqual(qc.qasm(), expected_qasm)
